Add check_translations script to report missing translation keys
New script to check which translation keys are missing compared to
en.json, without requiring regeneration. Also available via gen_strings
--show-missing flag for use after generation.

Usage:
  uv run python -m scripts.check_translations [lang]
  uv run python -m scripts.gen_strings --show-missing [lang]
